BraveWitch - early game creation 0.09 (fight options and console)

+++ src/main/java/AbleToAttack.java
@@ -1,5 +1,5 @@
 public interface AbleToAttack {
 
-double attack();
+int attack();
 
 }
